add optional 'align' parameter to addParagraph
Without this change it's very hard to center-align a paragraph that spans multiple lines. The module philosophy is to account for alignment in hPos (calculated before calling the method). Anyway this leads to troubles when trying to print a center-aligned paragraph that spans multiple lines: the first line starts at the right position, the following lines start aligned on the left